Move repositories to Entity namespace
