+++ src/OpenCATS/Entity/CompanyRepositoryException.php
